refactor: refactored user components from class to function
