+++ src/components/UserItem.tsx
@@ -1,34 +1,41 @@
-import React, { Component } from 'react';
+import React from 'react';
 
-class UserItem extends Component<UserProps, object> {
-  render() {
-    return (
-      <>
-        <div className="user">
-          <div className="user__title user__item">
-            <p>Name:</p>
-            {this.props.name}
-          </div>
-          <div className="user__file user__item">
-            <p>File:</p>
-            <img className="user__img" src={this.props.avatar} alt="image" />
-          </div>
-          <div className="user__birth user__item">
-            <p>Data:</p>
-            {this.props.birth}
-          </div>
-          <div className="user__gender user__item">
-            <p>Gender:</p>
-            {this.props.gender}
-          </div>
-          <div className="user__coffee user__item">
-            <p>Coffee:</p>
-            Somebody {this.props.doesUserLikeCoffe ? 'wants' : 'doesnt want'} coffee
-          </div>
+type UserProps = {
+  name: string;
+  birth: string;
+  gender: string;
+  doesUserLikeCoffe: boolean;
+  avatar: string;
+};
+
+function UserItem(props: UserProps) {
+  const { name, birth, gender, doesUserLikeCoffe, avatar } = props;
+  return (
+    <>
+      <div className="user">
+        <div className="user__title user__item">
+          <p>Name:</p>
+          {name}
         </div>
-      </>
-    );
-  }
+        <div className="user__file user__item">
+          <p>File:</p>
+          <img className="user__img" src={avatar} alt="image" />
+        </div>
+        <div className="user__birth user__item">
+          <p>Data:</p>
+          {birth}
+        </div>
+        <div className="user__gender user__item">
+          <p>Gender:</p>
+          {gender}
+        </div>
+        <div className="user__coffee user__item">
+          <p>Coffee:</p>
+          Somebody {doesUserLikeCoffe ? 'wants' : 'doesnt want'} coffee
+        </div>
+      </div>
+    </>
+  );
 }
 
 export default UserItem;
